+++ .envrc
@@ -1,0 +1,1 @@
+mamba activate pymultifracs
