+++ env.yml
@@ -13,4 +13,4 @@ dependencies:
   - pip
   - pip:
     - recombinator
-    - "--editable=git+https://github.com/neurospin/pymultifracs.git@master#egg=pymultifracs"
+    - pymultifracs
