+++ pymultifracs/jax/__init__.py
@@ -1,0 +1,6 @@
+from .utils import _binom
+from .wavelet import integrate_wt, _compute_leaders_jax, _correct_pleaders_jax
+from .scalingfun import _compute_Cmj, _compute_Cm_jax
+from .estimation import compute_etap
+from .mfa import _compute_weights, _compute_cm_jax
+from .simul import _gaussian_cme, _skewness_conv
