+++ pymultifracs/jax/jax.py
@@ -1,0 +1,3 @@
+import numpy as np
+import jax.numpy as jnp
+from jax.scipy.special import gammaln
