+++ pytest.ini
@@ -12,4 +12,5 @@ markers =
     viz
     robust
     benchmark
+    jax
 addopts = -nauto --cache-clear
